backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pgvector==0.2.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pydantic==2.5.2
langchain==0.0.350
openai==1.3.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.3
httpx==0.25.2 